+++ two_step_techniques/negativos_fiables.py
@@ -9,7 +9,7 @@ from two_step_techniques.metricas_distancia import comprobar_nombre_metrica, obt
 # Método de Rocchio. En cada método se utiliza un dataset ya convertido en PU (engineered):
 def rocchio(X, y, nombre_metrica):
     # Transformamos "X" e "y" en arrays de numpy para una mayor eficiencia:
-    X = np.asarray(X)
+    X = np.asarray(X, dtype = np.float64)
     y = np.asarray(y)
 
     # Comprobamos que la métrica sea válida:
@@ -49,7 +49,7 @@ def rocchio(X, y, nombre_metrica):
 # KNN (K-Nearest Neighbours):
 def knn(X, y, nombre_metrica, k, porcentaje_rn):
     # Transformamos "X" e "y" en arrays de numpy para una mayor eficiencia:
-    X = np.asarray(X)
+    X = np.asarray(X, dtype = np.float64)
     y = np.asarray(y)
 
     # Comprobamos que la métrica sea válida:
@@ -91,7 +91,7 @@ def knn(X, y, nombre_metrica, k, porcentaje_rn):
 # K-Means:
 def kmeans(X, y, nombre_metrica, k, porcentaje_rn, semilla = 1):
     # Transformamos "X" e "y" en arrays de numpy para una mayor eficiencia:
-    X = np.asarray(X)
+    X = np.asarray(X, dtype = np.float64)
     y = np.asarray(y)
 
     # Comprobamos que la métrica sea válida y obtenemos una métrica de scipy:
@@ -167,7 +167,7 @@ def kmeans(X, y, nombre_metrica, k, porcentaje_rn, semilla = 1):
 # K-Medoids:
 def kmedoids(X, y, nombre_metrica, k, porcentaje_rn, semilla = 1):
     # Transformamos "X" e "y" en arrays de numpy para una mayor eficiencia:
-    X = np.asarray(X)
+    X = np.asarray(X, dtype = np.float64)
     y = np.asarray(y)
 
     # Comprobamos que la métrica sea válida. Además, obtenemos una métrica de scipy:
@@ -245,7 +245,7 @@ def kmedoids(X, y, nombre_metrica, k, porcentaje_rn, semilla = 1):
 # CRNE:
 def crne(X, y, nombre_metrica, k, semilla = 1):
     # Transformamos "X" e "y" en arrays de numpy para una mayor eficiencia:
-    X = np.asarray(X)
+    X = np.asarray(X, dtype = np.float64)
     y = np.asarray(y)
     
     # Comprobamos que la métrica sea válida y obtenemos una métrica de scipy:
